requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
discord.py==1.4.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pokebase==1.3.0
pylint==2.5.3
requests==2.24.0
six==1.15.0
tinydb==4.1.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
wrapt==1.12.1
yarl==1.5.1